Обновление конфигурации базы данных: изменение имени базы данных и пользователя, а также обновление хоста. Исправление типов данных в моделях и схемах: замена Float на Integer для идентификаторов вопросов. Обновление миграций и удаление устаревших файлов миграций.

+++ app/migrations/versions/6b6aabeb13de_remove_password_column.py
@@ -1,8 +1,8 @@
-"""remove_password_column
+"""Remove password column
 
 Revision ID: 6b6aabeb13de
-Revises: 1c6d7302b37f
-Create Date: 2025-03-05 00:58:15.513017
+Revises: 32a8c1b7a322
+Create Date: 2025-03-05 00:50:15.513017
 
 """
 from typing import Sequence, Union
@@ -13,16 +13,21 @@ import sqlalchemy as sa
 
 # revision identifiers, used by Alembic.
 revision: str = '6b6aabeb13de'
-down_revision: Union[str, None] = '1c6d7302b37f'
+down_revision: Union[str, None] = '32a8c1b7a322'
 branch_labels: Union[str, Sequence[str], None] = None
 depends_on: Union[str, Sequence[str], None] = None
 
 
 def upgrade() -> None:
-    # Удаляем колонку password
-    op.drop_column('users', 'password')
+    # Проверяем существование колонки password перед удалением
+    conn = op.get_bind()
+    inspector = sa.inspect(conn)
+    columns = [col['name'] for col in inspector.get_columns('users')]
+    
+    if 'password' in columns:
+        op.drop_column('users', 'password')
 
 
 def downgrade() -> None:
-    # Восстанавливаем колонку password
-    op.add_column('users', sa.Column('password', sa.String(), nullable=True))
+    # Добавляем обратно колонку password
+    op.add_column('users', sa.Column('password', sa.String(), nullable=False))
